+++ src/core/api/SimBriefAeroflyApi.ts
@@ -51,15 +51,9 @@ export class SimBriefAeroflyApi extends SimBriefApi {
             );
         }
 
-        const { aeroflyAircraftCode, aeroflyAircraftLivery } = this.findAeroflyAircraftCode(
-            simbriefPayload.aircraft.icaocode,
-            simbriefPayload.general.icao_airline,
-        );
-
         const originRunwayOrientation = Number(simbriefPayload.origin.plan_rwy.replace(/\D+/, "")) * 10;
         const destinationRunwayOrientation = Number(simbriefPayload.destination.plan_rwy.replace(/\D+/, "")) * 10;
 
-        flight.aircraft = new AeroflySettingsAircraft(aeroflyAircraftCode, aeroflyAircraftLivery);
         flight.flightSetting = AeroflySettingsFlight.createInFeet(
             Number(simbriefPayload.origin.pos_long),
             Number(simbriefPayload.origin.pos_lat),
@@ -76,12 +70,25 @@ export class SimBriefAeroflyApi extends SimBriefApi {
                 runway: simbriefPayload.origin.plan_rwy,
             },
         );
-        flight.fuelLoadSetting = new AeroflySettingsFuelLoad(
-            aeroflyAircraftCode,
-            Number(simbriefPayload.fuel.plan_ramp),
-            Number(simbriefPayload.weights.payload),
-            "Keep",
-        );
+
+        try {
+            const { aeroflyAircraftCode, aeroflyAircraftLivery } = this.findAeroflyAircraftCode(
+                simbriefPayload.aircraft.icaocode,
+                simbriefPayload.general.icao_airline,
+            );
+            flight.aircraft = new AeroflySettingsAircraft(aeroflyAircraftCode, aeroflyAircraftLivery);
+            flight.fuelLoadSetting = new AeroflySettingsFuelLoad(
+                aeroflyAircraftCode,
+                Number(simbriefPayload.fuel.plan_ramp),
+                Number(simbriefPayload.weights.payload),
+                "Keep",
+            );
+        } catch (e) {
+            if (!(e instanceof Error)) {
+                throw e;
+            }
+        }
+
         flight.timeUtc = new AeroflyTimeUtc(new Date(simbriefPayload.times.sched_out));
 
         const waypoints = this.getWaypointsFromNavlog(simbriefPayload);