+++ types/core/converter/other/MsfsPlnToAeroflyFlightConverter.test.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=MsfsPlnToAeroflyFlightConverter.test.d.ts.map