+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## [2.2.0]
+## [2.2.1] - 2026-07-29
+
+- Improved handling of import files
+- Improved handling of missing aircraft matches from SimBrief
+
+## [2.2.0] - 2026-07-28
 
 - Updated `@modelcontextprotocol/sdk` to fix security advisory